Move VtoProductData adapters to fitting-room-data, breaking the cycle
The previous commit merged product-data.ts into product.ts, which forced
product.ts to import ResolvedFittingRoomItem from fitting-room-data.ts —
while fitting-room-data.ts already imports product.ts. Type-only, so no
runtime cycle, but still a layering violation: product is the general
layer, fitting-room composes products, so the dependency must point one
way (fitting-room -> product).

buildVtoProductDataFromResolved is an adapter that consumes
ResolvedFittingRoomItem — it belongs on the fitting-room side. Move it,
plus findRecommendedColorSize / findCsaByLabel (cohesive set, all
fitting-room-only callers), from product.ts into fitting-room-data.ts.

product.ts keeps the shared VtoProductData / VtoSizeData /
VtoSizeColorData types (no fitting-room dependency) and now imports
nothing from fitting-room-data. Dependency direction is one-way again.

Pure relocation — function bodies unchanged. check + build pass.
