+++ src/components/overlays/fitting-room/detail-accordion-item.tsx
@@ -6,9 +6,9 @@ import { ItemFitText } from '@/components/item-fit-text'
 import { LinkT } from '@/components/link'
 import { SizeSelector } from '@/components/size-selector'
 import { Text, TextT } from '@/components/text'
-import { ResolvedFittingRoomItem } from '@/lib/fitting-room-data'
+import { buildVtoProductDataFromResolved, ResolvedFittingRoomItem } from '@/lib/fitting-room-data'
 import { useTranslation } from '@/lib/locale'
-import { buildVtoProductDataFromResolved, VtoProductData } from '@/lib/product'
+import { VtoProductData } from '@/lib/product'
 import { useCss } from '@/lib/theme'
 import { Chevron } from './chevron'
